Ensure email address is unique
Additionally, remove the user view from the response body of the registration. User has to log in to reach this information, in addition it complicates things in the event that the email is already in use because the user can then infer whether the account exists or not which is a security risk.